WIP > move resuable invoke contract function to utils
